Python/pyFollower/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
deta==1.1.0
fake-useragent==0.1.11
fastapi==0.78.0
Flask==2.1.2
frozenlist==1.3.0
h11==0.13.0
httptools==0.4.0
idna==2.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
multidict==6.0.2
parse==1.18.0
pydantic==1.9.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-multipart==0.0.5
python-slugify==3.0.2
requests==2.25.1
requests-html==0.10.0
requests-toolbelt==0.9.1
sanic-routing==22.3.0
six==1.15.0
slugify==0.0.1
sniffio==1.2.0
soupsieve==2.1
starlette==0.19.1
text-unidecode==1.2
tqdm==4.55.1
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.2
uvicorn==0.17.6
uvloop==0.16.0
w3lib==1.22.0
websockets==10.3
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
